services/incidents/requirements.txt
PyJWT==2.8.0
boto3==1.34.0
python-ulid==2.2.0
